# frontend/openapi/openapi.json
{
  "components" : {
    "schemas" : {
...
      "ChallengeRequest" : {
        "properties" : {
          "challenge" : {
            "type" : "string"
          },
          "code" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "CodeRequest" : {
        "properties" : {
          "code" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "Enrollment" : {
        "properties" : {
          "otpauthUri" : {
            "type" : "string"
          },
          "secret" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "RecoveryCodeList" : {
        "properties" : {
          "recoveryCodes" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
...
      "StepUpResponse" : {
        "properties" : {
          "accessToken" : {
            "type" : "string"
          },
          "accessTokenExpiresAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "mfaAt" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "TokenResponse" : {
        "properties" : {
          "accessToken" : {
            "type" : "string"
          },
          "accessTokenExpiresAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "challenge" : {
            "type" : "string"
          },
          "challengeExpiresAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "permissions" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "refreshToken" : {
            "type" : "string"
          },
          "refreshTokenExpiresAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "roles" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "status" : {
            "enum" : [ "SIGNED_IN", "MFA_REQUIRED", "MFA_ENROLLMENT_REQUIRED" ],
            "type" : "string"
          },
          "tokenType" : {
            "type" : "string"
          },
          "userId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    }
  },
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/audit/operations" : {
      "get" : {
        "operationId" : "operations",
        "parameters" : [ {
          "in" : "query",
          "name" : "actorId",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "from",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "to",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "processName",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "entityType",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "entityId",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "offset",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "additionalProperties" : { },
                  "type" : "object"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "audit-controller" ]
      }
    },
...
    "/api/auth/challenge/enroll" : {
      "post" : {
        "operationId" : "enrollUnderChallenge",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ChallengeRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Enrollment"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "mfa-controller" ]
      }
    },
    "/api/auth/challenge/enroll/confirm" : {
      "post" : {
        "operationId" : "confirmUnderChallenge",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ChallengeRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RecoveryCodeList"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "mfa-controller" ]
      }
    },
    "/api/auth/challenge/verify" : {
      "post" : {
        "operationId" : "verify_1",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ChallengeRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TokenResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "mfa-controller" ]
      }
    },
...
    "/api/auth/mfa/enroll/confirm" : {
      "post" : {
        "operationId" : "confirm",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CodeRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RecoveryCodeList"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "mfa-controller" ]
      }
    },
...
    "/api/auth/step-up" : {
      "post" : {
        "operationId" : "stepUp",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CodeRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StepUpResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "mfa-controller" ]
      }
    },
...
  }
}